backend/requirements.txt
# Hanco-AI Backend Dependencies
# Python 3.11+ (3.11.x recommended for production)

# ==================== Python ====================
# Python 3.11.x (recommended)

# ==================== FastAPI Framework ====================
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.4
pydantic-settings==2.2.1
email-validator==2.1.1

# ==================== Firebase ====================
firebase-admin==6.5.0
google-cloud-firestore==2.15.0

# ==================== AI Services ====================
google-generativeai==0.3.2
openai==1.52.0

# ==================== ML / ONNX ====================
numpy==1.26.4
onnxruntime==1.17.3
onnx==1.15.0
lightgbm==4.3.0
onnxmltools==1.12.0

# ==================== Web Scraping ====================
# crawl4ai==0.7.2  # Temporarily disabled due to dependency conflicts
beautifulsoup4==4.12.3
lxml==5.3.0
playwright==1.42.0

# ==================== HTTP Clients ====================
httpx==0.27.2
aiohttp==3.11.11
requests==2.31.0

# ==================== Data Processing ====================
pandas==2.1.4

# ==================== Security / Auth ====================
python-dotenv==1.0.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
slowapi==0.1.9

# ==================== ML Training ====================
scikit-learn==1.4.2
skl2onnx==1.16.0

# ==================== Scheduler ====================
APScheduler==3.10.4
pytz==2024.1

# ==================== Testing ====================
pytest==7.4.4
pytest-asyncio==0.23.6

# ==================== Development ====================
black==24.4.2
ruff==0.3.7
mypy==1.10.0
